Backend/requirements.txt
asgiref==3.3.1
certifi==2019.9.11
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-filter==1.1.0
django-js-asset==1.2.2
djangorestframework==3.12.2
drf-yasg==1.20.0
easy-thumbnails==2.7.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
lark-parser==0.8.2
MarkupSafe==1.1.1
packaging==20.9
Pillow==8.1.2
pipenv==2018.11.26
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==16.7.7
virtualenv-clone==0.5.3
